+++ app/views/movies/index.html.erb
@@ -1,5 +1,5 @@
 <div class="text-center my-4">
-  <h1><%= switch_movies_title %></h1>  <%# ===== `movies_helper:17`参照 ===== %>
+  <h1><%= switch_page_title %>動画教材</h1>  <%# ===== `movies_helper:17`参照 ===== %>
 </div>
 
 <div class="container-fluid">
