+++ app/views/texts/index.html.erb
@@ -1,5 +1,5 @@
 <h1 class = "text-center mt-2 mb-4">
-<%= select_pagetitle %>
+<%= switch_page_title %>
 <br class = "d-sm-none">テキスト教材
 </h1> 
 <div class="row">
